오류수정 #09

+++ Jenkinsfile
@@ -31,15 +31,14 @@ data = json.load(open("env.json"))
 with open(".env.local", "w") as f:
     for k, v in data.items():
         f.write(f"{k}={v}\\n")'
-                        rm env.json
-                        '''
+                        ''' && rm env.json
                     }
-                    sh '''
-                    docker build -t mantal-map-frontend:${BUILD_TIME} .
-                    aws ecr get-login-password --region $AWS_REGION | docker login --username AWS --password-stdin $ECR_URL
-                    docker tag mantal-map-frontend:${BUILD_TIME} $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME}
+                    sh """
+                    docker build -t mantal-map-frontend:${BUILD_TIME} . && \
+                    aws ecr get-login-password --region $AWS_REGION | docker login --username AWS --password-stdin $ECR_URL && \
+                    docker tag mantal-map-frontend:${BUILD_TIME} $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME} && \
                     docker push $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME}
-                    '''
+                    """
                 }
             }
         }
@@ -61,14 +60,13 @@ data = json.load(open("env.json"))
 with open(".env.${DEPLOY_ENV}", "w") as f:
     for k, v in data.items():
         f.write(f"{k}={v}\\n")'
-                            rm env.json
-                            """
+                            """ && rm env.json
                         }
 
                         sh """
-                        docker build -t mantal-map-backend:${BUILD_TIME} .
-                        aws ecr get-login-password --region $AWS_REGION | docker login --username AWS --password-stdin $ECR_URL
-                        docker tag mantal-map-backend:${BUILD_TIME} $ECR_URL:backend-${DEPLOY_ENV}-${BUILD_TIME}
+                        docker build -t mantal-map-backend:${BUILD_TIME} . && \
+                        aws ecr get-login-password --region $AWS_REGION | docker login --username AWS --password-stdin $ECR_URL && \
+                        docker tag mantal-map-backend:${BUILD_TIME} $ECR_URL:backend-${DEPLOY_ENV}-${BUILD_TIME} && \
                         docker push $ECR_URL:backend-${DEPLOY_ENV}-${BUILD_TIME}
                         """
                     }
@@ -93,28 +91,28 @@ with open(".env.${DEPLOY_ENV}", "w") as f:
 
                     if (params.DEPLOY_TARGET == 'frontend' || params.DEPLOY_TARGET == 'both') {
                         sh """
-                        scp -i $SSH_KEY frontend/${FRONT_ENV_FILE} ubuntu@${TARGET_SERVER}:/home/ubuntu/
-                        ssh -i $SSH_KEY -o StrictHostKeyChecking=no ubuntu@${TARGET_SERVER} \"
-                          docker pull $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME}
-                          docker stop mantal-map-frontend || true
-                          docker rm mantal-map-frontend || true
+                        scp -i $SSH_KEY frontend/${FRONT_ENV_FILE} ubuntu@${TARGET_SERVER}:/home/ubuntu/ && \
+                        ssh -i $SSH_KEY -o StrictHostKeyChecking=no ubuntu@${TARGET_SERVER} "\
+                          docker pull $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME} && \
+                          (docker stop mantal-map-frontend || true) && \
+                          (docker rm mantal-map-frontend || true) && \
                           docker run -d --name mantal-map-frontend --restart unless-stopped -p 8003:3000 --env-file /home/ubuntu/${FRONT_ENV_FILE} $ECR_URL:frontend-${params.DEPLOY_ENV}-${BUILD_TIME}
-                        \"
+                        "
                         """
                     }
 
                     if (params.DEPLOY_TARGET == 'backend' || params.DEPLOY_TARGET == 'both') {
                         sh """
-                        scp -i $SSH_KEY backend/${BACK_ENV_FILE} ubuntu@${TARGET_SERVER}:/home/ubuntu/
-                        ssh -i $SSH_KEY -o StrictHostKeyChecking=no ubuntu@${TARGET_SERVER} \"
-                          docker pull $ECR_URL:backend-${params.DEPLOY_ENV}-${BUILD_TIME}
-                          docker stop mantal-map-backend || true
-                          docker rm mantal-map-backend || true
+                        scp -i $SSH_KEY backend/${BACK_ENV_FILE} ubuntu@${TARGET_SERVER}:/home/ubuntu/ && \
+                        ssh -i $SSH_KEY -o StrictHostKeyChecking=no ubuntu@${TARGET_SERVER} "\
+                          docker pull $ECR_URL:backend-${params.DEPLOY_ENV}-${BUILD_TIME} && \
+                          (docker stop mantal-map-backend || true) && \
+                          (docker rm mantal-map-backend || true) && \
                           docker run -d --name mantal-map-backend --restart unless-stopped -p 8002:8000 \\
                             --env ENVIRONMENT=${params.DEPLOY_ENV} \\
                             --env-file /home/ubuntu/${BACK_ENV_FILE} \\
                             $ECR_URL:backend-${params.DEPLOY_ENV}-${BUILD_TIME}
-                        \"
+                        "
                         """
                     }
                 }
